refactor: change application name on created files
