+++ docs/infrastructure.md
@@ -1,3 +1,3 @@
-# Mintsu's Infrastructure
+# Meent's Infrastructure
 
 WIP.
