+++ src/test/kotlin/com/meent/app/DevApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.mintsu.dev
+package com.meent.app
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
